+++ src/components/DualListBox/internal/SelectedList/SelectedList.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 import * as Styled from "./styled";
-import { SelectedItem } from "../../DualListBox";
+import { DualListBoxItem } from "../../DualListBox";
 import Icon from "../../../Icon/Icon";
 import { useTheme } from "../../../../themes";
 import Typography from "../../../Typography/Typography";
-import ToggleButton from "../../../ToggleButton/ToggleButton";
 
 export const SelectedList: React.FunctionComponent<{
-  items: SelectedItem[];
-  onRemove?: (id: string) => void;
-  onToggleChange?: (id: string) => void;
-}> = ({ items, onRemove, onToggleChange }) => {
+  items: DualListBoxItem[];
+  selectedItemTitle: string;
+  onRemove?: (item: DualListBoxItem) => void;
+  onToggleChange?: (item: DualListBoxItem) => void;
+}> = ({ items, selectedItemTitle, onRemove }) => {
   const theme = useTheme();
 
   return (
@@ -18,9 +18,9 @@ export const SelectedList: React.FunctionComponent<{
       <Styled.SelectedHeader>
         <Typography size="sm" weight="bold">
           <Typography color="primary" component="span" size="sm" weight="bold">
-            {items.length}個
+            {items.length}{" "}
           </Typography>
-          選択済み
+          {selectedItemTitle}
         </Typography>
       </Styled.SelectedHeader>
       <Styled.SelectedList>
@@ -29,21 +29,12 @@ export const SelectedList: React.FunctionComponent<{
             key={item.id}
             isLastIndex={i + 1 === items.length}
           >
-            <Typography>{item.label}</Typography>
+            {item.content}
             <Styled.Action>
-              {onToggleChange && (
-                <ToggleButton
-                  checkedText=""
-                  unCheckedText=""
-                  width="48px"
-                  checked={item.checked}
-                  onChange={() => onToggleChange(item.id)}
-                />
-              )}
               {onRemove && (
                 <Styled.RemoveButton
                   type={"button"}
-                  onClick={() => onRemove(item.id)}
+                  onClick={() => onRemove(item)}
                 >
                   <Icon name="close" size="sm" color={theme.palette.black} />
                 </Styled.RemoveButton>