+++ src/components/WeekTime/internal/WeekTimeElement/constants.ts
@@ -1,0 +1,2 @@
+// 0 to 24 array
+export const timeList = Array.from(Array(24)).map((_, i) => i);